Updating curator and checkstyle.
Updating Curator had some unanticipated side effects due to issues that were addressed
in CURATOR-136 ( https://github.com/apache/curator/commit/05254afea628a1018e1058d230a568746f98df5a#diff-987cef79524e0a6cb118d2df685f041d ).

The first issue is that by validating the path earlier in the distributed double barrier it caused it throw an IAE instead of a NPE
when the path was null. Since Cultivar consistently uses checkNotNull to check for null in these cases the fix was to add checkNotNull
into the factory.

The second issue is that, previously, `framework.newNamespaceAwareEnsurePath("test")` would work if framework was a CuratorFramework with
a null namespace. That is no longer the case, requiring it to be "/test" instead.  This also changed the response from getPath that was then
called on the newNamespaceAwareEnsurePath.